Fixed some reading history bugs
The check for the current page on a volume wasn't selecting the row correctly so any volume update would add a reading history row. The 'clean' states in the VolumeEditor were also incorrect, they'd behind the non-clean states but were basically the same and it'd cause an always dirty state where it'd alternate between the last two page values.