Added a link to / in the header to exit out of a series/volume
Centered everything on the page so when the volume editor isn't visible the page doesn't look awkwardly unbalanced. However after a volume has been selected for the first time it stops trying to center things, so when the user selects a different series the selectors don't jump to the center of the page again.