Refactored to the server architecture
I tried to go for Clean Architecture, or at least what I currently understand of it and to the degree it felt made sense to do.

There are now three layers.

First, the "controllers" layer which are the http endpoint handlers. They handle converting an http request into a call to an interactor, and converting data to/from the entities it uses, responding with its own version of the relevant data which lets the controllers get rid of data that's unnecessary/unwanted, and to keep the entity structs the database operates on completely decoupled from anything like json marshalling.

Second, the "interactors" layer, which is a proxy between controllers and repositories, and handles business logic concerns (eventually that will include validation).

Third, the "repositories" layer, which is just a layer to encapsulate storing data in the database. It returns entities whenever needed. Both the interactors and repositories deal in entities.

Packages are a bit more concise, all controllers/repositories/interactors live within a shared package for that layer instead of a separate package for each of series/volume/readinghistory within each layer.

All of this is just to try and make concerns as separate and encapsulated as possible at their level.