Replaced a workaround for hiding the menu popup until it could be positioned
Popper is used to put the menu popup in the right spot when the '...' button next to a series or volume is clicked, but initializing it in a normal `useEffect` hook you'd sometimes see a frame of the menu interfering with the layout of the list. I had worked around this by adding some tailwind classes to position it way off screen until that `useEffect` hook could put it in the right spot. I switched it to `useLayoutHook` so it is synchronously positioned after it shows, no need for the off screen workaround.