Color changes and responsive view
For viewports under the `md` breakpoint the app will stack.

The app now has a max width so the reading history and "Edit" button aren't so far to the right that they're hard to notice after focusing on the left side of the page.

Styles on the selectors (hover and selected states) are now more obvious. Just the blue text color was kind of hard to see at a glance.

There are a few other tweaks to improve the layout on both desktop and mobile, but most are super worth explaining individual.