Added explicit "read pages" tracking and volume status
The amount of pages read was dynamic and frontend only previously. It would be calculated by getting the difference between consecutive reading history items, so deleting a reading history item would change the pages read that were displayed in the side bar. This made it very difficult to do anything beyond that side bar. I would like to be able to show a reading graph or some sort of analytics to users at some point, but that was impossible as it was. It has been refactored to prompt for a number of pages read (defaulting to the delta) whenever the user edits a volume's current page, and that page count is stored in the reading history table now alongside the current page that was tracked. This way analytics can be easily done.

I also added a status to each volume. The statuses are: planning, reading, completed, dropped. There's a new selector (non-editable) that shows up above the series selector that only shows volumes in that reading status, so if a user just bookmarks `/` they won't have to dig to find the volume they're on. The statuses are shown as a little circle of color next to the volume name in the volume selector and in the header on the volume editor so it's again easy to see what volume you're on.